move repair step to stable12
because we decided to backport it the repair step needs to be executed
already on stable12
